move bitsize_from_type_token() to shared.rs
